API: Call them shortcuts. And fix more renamings missed.
