+++ bluesky/standard_config.py
@@ -14,7 +14,7 @@ from bluesky.hardware_checklist import (connect_mds_mongodb,
                                         assert_pv_equal, assert_pv_greater,
                                         assert_pv_less, assert_pv_in_band,
                                         assert_pv_out_of_band)
-import bluesky.brittle_scans as bs
+import bluesky.shortcuts as sc
 
 logger = logging.getLogger(__name__)
 
@@ -54,21 +54,21 @@ except KeyError:
 register_mds(RE)  # subscribes to MDS-related callbacks
 
 
-# Instantiate brittle scans.
-ct = count = bs.Count(RE)
-ascan = bs.AbsoluteScan(RE)
-mesh = bs.OuterProductAbsoluteScan(RE)
-a2scan = a3scan = bs.InnerProductAbsoluteScan(RE)
-dscan = lup = bs.DeltaScan(RE)
-d2scan = d3scan = bs.InnerProductDeltaScan(RE)
-th2th = bs.ThetaTwoThetaScan(RE)
-hscan = bs.HScan(RE)
-kscan = bs.KScan(RE)
-lscan = bs.LScan(RE)
-tscan = bs.AbsoluteTemperatureScan(RE)
-dtscan = bs.DeltaTemperatureScan(RE)
-hklscan = bs.OuterProductHKLScan(RE)
-hklmesh = bs.InnerProductHKLScan(RE)
+# Instantiate shortcuts.
+ct = count = sc.Count(RE)
+ascan = sc.AbsoluteScan(RE)
+mesh = sc.OuterProductAbsoluteScan(RE)
+a2scan = a3scan = sc.InnerProductAbsoluteScan(RE)
+dscan = lup = sc.DeltaScan(RE)
+d2scan = d3scan = sc.InnerProductDeltaScan(RE)
+th2th = sc.ThetaTwoThetaScan(RE)
+hscan = sc.HScan(RE)
+kscan = sc.KScan(RE)
+lscan = sc.LScan(RE)
+tscan = sc.AbsoluteTemperatureScan(RE)
+dtscan = sc.DeltaTemperatureScan(RE)
+hklscan = sc.OuterProductHKLScan(RE)
+hklmesh = sc.InnerProductHKLScan(RE)
 
 
 def olog_wrapper(logbook, logbooks):
